// MySchool-Microservices/frontend/src/app/core/models/students.model.ts
export interface PagedResultDto<T> {
  data: T[];
  pageNumber: number;
  pageSize: number;
  totalCount: number;
  totalPages: number;
}